pyproject.toml
[tool.poetry]
name = "microblog"
version = "0.1.0"
description = "This project give me job"
authors = ["Miron <miron.nicolaevich@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.32"}
fastapi = "^0.112.2"
uvicorn = "^0.30.6"
pydantic-settings = "^2.4.0"
asyncpg = "^0.29.0"
alembic = "^1.13.2"
greenlet = "^3.0.3"
gunicorn = "^23.0.0"
pydantic = {extras = ["email"], version = "^2.8.2"}
pyjwt = {extras = ["crypto"], version = "^2.9.0"}
bcrypt = "^4.2.0"
python-multipart = "^0.0.10"
redis = {extras = ["hiredis"], version = "^5.1.1"}


[tool.poetry.group.dev.dependencies]
mypy = "^1.11.1"
black = "^24.8.0"
isort = "^5.13.2"
flake8 = "^7.1.1"
pre-commit = "^3.8.0"
pytest = "^8.3.2"
coverage = "^7.6.1"
pytest-cov = "^5.0.0"
factory-boy = "^3.3.1"
faker = "^27.4.0"
pytest-asyncio = "^0.24.0"
asgi-lifespan = "^2.1.0"
httpx = "^0.27.2"
types-redis = "^4.6.0.20241004"


[tool.coverage.run]
branch = true
source = ["src"]


[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.black]
line-length = 79
include = '\.py'
target-version = ['py312']

[tool.isort]
profile = "black"
line_length = 79
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
